fix: provide non-android pip stub
